moved some classes from `c.e.h_f.ui.domain.weight` to `c.e.h_f.model.weight`
