+++ app/src/main/java/com/example/happ_frontend/ui/screens/weight/AddEventDialog.kt
@@ -1,6 +1,5 @@
 package com.example.happ_frontend.ui.screens.weight
 
-import AuthFormNumberField
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,19 +12,13 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.happ_frontend.R
-import com.example.happ_frontend.ui.domain.weight.WeightCalendarEvent
-import com.example.happ_frontend.ui.domain.weight.WeightHistoryViewModel
 
 /**
  * A dialog for adding a new weight event to the calendar.
